:art: Rename files and stub CaptionPane
Ref #3
