+++ Subtend/src/GUI/CaptionPane.cpp
@@ -1,0 +1,12 @@
+#include "sbtpch.h"
+#include "CaptionPane.h"
+
+namespace Subtend
+{
+
+    void Subtend::CaptionPane::RenderUI()
+    {
+    
+    }
+
+}
