+++ Subtend/src/GUI/CaptionPane.h
@@ -1,0 +1,13 @@
+#pragma once
+
+namespace Subtend
+{
+    class CaptionPane
+    {
+    public:
+        CaptionPane() = default;
+
+        void RenderUI();
+    };
+}
+
